progress

+++ frontend/src/components/EditTeam.vue
@@ -3,7 +3,7 @@
         <div class="info">
             <h2>Editar Equipo</h2>
             <h5>Ingresa los campos para actualizar equipo</h5>
-
+            <div class="form">
                 <b-form @submit.prevent="editTeam()">
 
                     <b-form-group id="input-group-1" label="Nombre del equipo:" label-for="input-1">
@@ -26,10 +26,11 @@
                         </option>
                         </b-form-select>
                     </b-form-group>
-                    <b-button type="submit" variant="primary">Enviar</b-button>
+                    <b-button type="submit" variant="primary">Editar Equipo</b-button>
                     <b-button @click="deleteTeam()" variant="danger">Eliminar equipo</b-button>
 
                 </b-form>
+            </div>
         </div>
     </div>
 </template>
@@ -42,7 +43,7 @@ export default {
 
     data() {
         return {
-            id: null,
+            id: '',
             name: '',
             manager: '',
             team: '',
@@ -55,6 +56,7 @@ export default {
             const data = { name: this.name, manager_name: this.manager.id, league_name: this.team.league_name }
             const path = 'http://127.0.0.1:8000/api/teams/edit/' + this.id
             await axios.put(path, data) 
+            this.goToTeam()
         },
         async getData () {
             this.id = this.$route.params.id
@@ -77,9 +79,23 @@ export default {
             this.name = this.team.name
         },
         async deleteTeam() {
-            this.id = this.$route.params.id
-            const path = 'http://127.0.0.1:8000/api/teams/delete/' + this.id
-            await axios.delete(path)
+            let text = "¿Seguro deseas eliminar este equipo?";
+            if (confirm(text) == true) {
+                this.id = this.$route.params.id
+                const path = 'http://127.0.0.1:8000/api/teams/delete/' + this.id
+                await axios.delete(path)
+                this.goToLeague()
+            } else {
+
+            }
+        },
+        goToTeam() {
+            alert("Equipo actualizado correctamente")
+            this.$router.back() 
+        },
+        goToLeague() {
+            alert("Equipo borrado correctamente")
+            this.$router.push("/leagues")           
         }
     },
     created() {
@@ -92,6 +108,9 @@ export default {
 .info {
     text-align: left;
     margin: 15px;
+}
+.form {
+    margin: 25px 50px 0 50px;
 }
 h2 {
     border-bottom: 1px solid black;
